+++ src/utils/adopter.js
@@ -53,6 +53,10 @@ export function adopt (node) {
   // make sure a node isn't already adopted
   if (node.instance instanceof Base) return node.instance
 
+  if (node.nodeName === '#document-fragment') {
+    return new elements.Fragment(node)
+  }
+
   // initialize variables
   var className = capitalize(node.nodeName || 'Dom')